Points: move describeFeatureRowState out of the panel for clean Fast Refresh
PointsFeatureFilterPanel.tsx exported both the component and
describeFeatureRowState (a runtime function). That single non-component
runtime export disqualifies the module from Vite React Fast Refresh, so
every edit did a full reload and dropped React state (the selected
layer). Move the classifier + featureRowOpacity + their types to a
sibling featureRowState.ts; the panel and the test import from there.

(Only the runtime function had to move — type exports are erased and
don't affect Fast Refresh; the types moved with it for cohesion / to
avoid a circular import.)
